requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
idna==3.10
joblib==1.4.2
kafka-python==2.0.2
kafka-python-ng==2.2.3
numpy==2.2.0
pandas==2.2.3
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
